+++ server/src/main/java/net/feedbacky/app/data/board/webhook/Webhook.java
@@ -64,7 +64,8 @@ public class Webhook implements Serializable {
     IDEA_COMMENT(2, "Idea Commented - `${idea.name}`"), IDEA_COMMENT_DELETE(3, "Idea Comment Deleted - `${idea.name}`"),
     IDEA_EDIT(4, "Idea Edited - `${idea.name}`"), IDEA_TAG_CHANGE(6, "Tags of Idea Changed - `${idea.name}`"),
     IDEA_OPEN(7, "Idea Opened - `${idea.name}`"), IDEA_CLOSE(8, "Idea Closed - `${idea.name}`"),
-    SAMPLE_EVENT(9, "I am alive! Sample request received.");
+    SAMPLE_EVENT(9, "I am alive! Sample request received."), IDEA_COMMENTS_RESTRICT(10, "Idea Commenting Restricted - `${idea.name}`"),
+    IDEA_COMMENTS_ALLOW(10, "Idea Commenting Allowed - `${idea.name}`");
 
     private final int id;
     private final String message;