+++ server/src/main/java/net/feedbacky/app/data/idea/dto/PatchIdeaDto.java
@@ -26,5 +26,6 @@ public class PatchIdeaDto {
   @Length(min = 20, max = 1800, message = "Description cannot be shorter than 20 or longer than 1800 characters.")
   private String description;
   private Boolean open;
+  private Boolean commentingRestricted;
 
 }